+++ node-api/src/test/kotlin/net/corda/nodeapi/internal/serialization/carpenter/CarpenterExceptionTests.kt
@@ -69,7 +69,7 @@ class CarpenterExceptionTests {
     }
 
     @Test
-    fun carpenterExcpRethrownAsNSE() {
+    fun carpenterExceptionRethrownAsNotSerializableException() {
         data class C2 (val i: Int) : TestInterface {
             override fun runThing() = 1
         }